requirements.txt
aiofiles==0.5.0
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
autopep8==1.5.4
bcrypt==3.2.0
blis==0.7.4
catalogue==2.0.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cymem==2.0.5
dnspython==2.0.0
email-validator==1.1.2
es-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.0.0/es_core_news_sm-3.0.0-py3-none-any.whl
fastapi==0.61.2
fastapi-users==5.0.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.0.4
h11==0.9.0
idna==2.10
importlib-metadata==3.4.0
itsdangerous==1.1.0
Jinja2==2.11.2
makefun==1.9.5
MarkupSafe==1.1.1
motor==2.3.0
murmurhash==1.0.5
numpy==1.20.0
orjson==3.4.3
packaging==20.9
passlib==1.7.4
pathy==0.3.5
preshed==3.0.5
promise==2.3
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.7.2
PyJWT==2.0.0
pymongo==3.11.1
pyparsing==2.4.7
python-decouple==3.3
python-multipart==0.0.5
PyYAML==5.3.1
requests==2.25.0
Rx==1.6.1
six==1.15.0
smart-open==3.0.0
spacy==3.0.1
spacy-legacy==3.0.1
srsly==2.4.0
starlette==0.13.6
thinc==8.0.1
toml==0.10.2
tqdm==4.56.0
typer==0.3.2
typing-extensions==3.7.4.3
ujson==3.2.0
urllib3==1.26.2
uvicorn==0.11.8
wasabi==0.8.2
websockets==8.1
zipp==3.4.0
